Category add/update/delete added

+++ src/Services/Assessment/src/Application/MainCategories/Commands/DeleteMainCategory/DeleteMainCategoryCommand.cs
@@ -21,12 +21,14 @@ public class DeleteMainCategoryCommand : IRequest<ApiResult<string>>
 
         public async Task<ApiResult<string>> Handle(DeleteMainCategoryCommand request, CancellationToken cancellationToken)
         {
-            var mainCategory = await _context.MainCategories.SingleOrDefaultAsync(x=>x.Id == request.Id && x.Deleted == null, cancellationToken: cancellationToken);
+            var mainCategory = await _context.MainCategories.SingleOrDefaultAsync(x => x.Id == request.Id && x.Deleted == null, cancellationToken: cancellationToken);
             if (mainCategory == null)
                 throw new NotFoundException("MainCategory not found", request.Id);
 
-            // todo check if main category has any sub categories
-            
+            var categories = await _context.Categories.Where(x => x.Id == request.Id && x.Deleted == null).ToListAsync(cancellationToken: cancellationToken);
+            if (categories.Any())
+                throw new BadRequestException("MainCategory has categories");
+
             _context.MainCategories.Remove(mainCategory);
 
             await _context.SaveChangesAsync(cancellationToken);
